Created Sync Process

+++ src/main/java/org/example/bronze/metadata/FileVersion.java
@@ -2,7 +2,7 @@ package org.example.bronze.metadata;
 
 import java.time.Instant;
 
-public record FileVersion(long version, boolean isFullSnapshot, String filePath,
+public record FileVersion(int version, boolean isFullSnapshot, String filePath,
                           String hash, long size, Instant createdAt)
 {
 }
